ref(ui): Split out modals, prevent api.tsx from importing everything
Importing certain files drags in a ton of the app with it. By splitting these two modals out of the modal file we can prevent a lot of the app from being imported from `app/api.tsx`

see the following chain where the file we wanted was `app/chartcuterie/config.tsx` but got tons of components
```
 @ ../node_modules/platformicons/svg/ sync ^\.\/.*\.svg$ ./HTML5.svg
 @ ../node_modules/platformicons/build/index.js 1:4490-4517
 @ ./app/data/forms/projectGeneralSettings.tsx 5:0-45 52:48-60
 @ ./app/data/forms/ sync \.tsx?$ ./projectGeneralSettings.tsx
 @ ./app/actionCreators/formSearch.tsx 33:18-67
 @ ./app/components/search/sources/formSource.tsx 3:0-65 82:4-17
 @ ./app/components/search/index.tsx 11:0-69 110:42-52
 @ ./app/components/modals/commandPalette.tsx
 @ ./app/actionCreators/modal.tsx 103:20-69
 -- > @ ./app/api.tsx 7:0-74 164:2-19 230:6-14
 @ ./app/utils/useApi.tsx
 @ ./app/utils/queryClient.tsx 4:0-41 31:14-20 119:14-20
 @ ./app/views/alerts/utils/migrationUi.ts
 @ ./app/views/alerts/wizard/options.tsx 9:0-85 140:22-50
 @ ./app/views/alerts/rules/metric/details/metricChartOption.tsx 17:0-75 305:16-37
 @ ./app/chartcuterie/metricAlert.tsx 5:0-137 71:8-33 100:8-33 103:22-54
 @ ./app/chartcuterie/config.tsx 13:0-50 32:0-25
```
