+++ static/app/stores/modalStore.tsx
@@ -1,6 +1,6 @@
 import {createStore} from 'reflux';
 
-import {ModalOptions, ModalRenderProps} from 'sentry/actionCreators/modal';
+import type {ModalOptions, ModalRenderProps} from 'sentry/actionCreators/modal';
 
 import {CommonStoreDefinition} from './types';
