+++ static/app/views/alerts/utils/migrationUi.ts
@@ -2,7 +2,7 @@ import {Organization} from 'sentry/types';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
 import {Dataset, MetricRule} from 'sentry/views/alerts/rules/metric/types';
-import {CombinedMetricIssueAlerts} from 'sentry/views/alerts/types';
+import type {CombinedMetricIssueAlerts} from 'sentry/views/alerts/types';
 
 // TODO(telemetry-experience): remove when the migration is complete
 export const hasMigrationFeatureFlag = (organization: Organization): boolean =>
